Backend/python-scraper/requirements.txt
beautifulsoup4
requests
pymongo
fastapi
uvicorn
python-dotenv
